move PEdge
