+++ src/main/java/micycle/pgs/utility/SpiralQuadrangulation.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-import micycle.pgs.PGS.PEdge;
 import processing.core.PVector;
 
 /**
